feat: reducing database fetch and improving response time

+++ backend/src/main/java/dev/cassiano/to_do_api/services/TaskService.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import dev.cassiano.to_do_api.entities.Task;
-import dev.cassiano.to_do_api.entities.User;
 import dev.cassiano.to_do_api.exceptions.customs.NotFoundException;
 import dev.cassiano.to_do_api.repositories.TaskRepository;
 
@@ -18,14 +17,17 @@ public class TaskService {
         return taskRepository.save(task);
     }
 
-    public Task getById(Long id, User user) throws NotFoundException{
-        return taskRepository.findByIdAndUser(id, user)
+    public Task getById(Long id, String user_email) throws NotFoundException{
+        Task task = taskRepository.findById(id)
             .orElseThrow(() -> new NotFoundException("This task doesn't exists"));
+        if(task.getUser().getEmail() != user_email) throw new NotFoundException("This task doesn't exists");
+        return task;
     }
 
-    public void deleteById(Long id, User user) throws NotFoundException{
-        taskRepository.findByIdAndUser(id, user)
+    public void deleteById(Long id, String user_email) throws NotFoundException{
+        Task task = taskRepository.findById(id)
             .orElseThrow(() -> new NotFoundException("This task doesn't exists"));
+        if(task.getUser().getEmail() != user_email) throw new NotFoundException("This task doesn't exists");
         taskRepository.deleteById(id);
     }
 }
